Fixed invalid word form
This fixes invalid word forms where parentheses are kept but inside tokens contains spaces I've came across this bug while trying to tokenize a text containing the following sequence "... sic (! ) ..." the three tokens joined together don't provide the real word form because it contains a space. The only way to get it right is to get a substring of the real input.